requirements.txt
geopy==2.4.1
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
